add critical test for  demonstration
