+++ orchestration/dags/data_gcp_dbt/models/analytics/enriched/enriched_venue_data.yml
@@ -11,87 +11,87 @@ models:
             field: venue_id
             config:
               severity: error
-              error_if: ">10"
-  - name: venue_public_name
-  - name: venue_name
-  - name: venue_booking_email
-  - name: venue_address
-  - name: venue_latitude
-  - name: venue_longitude
-  - name: venue_department_code
-  - name: venue_postal_code
-  - name: venue_city
-  - name: venue_siret
-  - name: venue_is_virtual
-  - name: venue_is_permanent
-  - name: venue_managing_offerer_id
-  - name: venue_creation_date
-  - name: offerer_name
-  - name: offerer_validation_status
-    description: "Status of validation of an offerer"      
-  - name : venue_type_label
-    description: Type of venue
-  - name: venue_label
-  - name: total_bookings
-    description: 'Total number of bookings generated by a venue'
-  - name: non_cancelled_individual_bookings
-  - name: non_cancelled_collective_bookings
-  - name: first_individual_booking_date
-    description: First booking date on an individual offer
-  - name: last_individual_booking_date
-    description: Last booking date on an individual offer
-  - name: first_collective_booking_date
-    description: First booking date on a collective offer
-  - name: last_collective_booking_date
-    description: Last booking date on a collective offer
-  - name: non_cancelled_bookings
-    description: Number of non cancelled bookings by the venue.
-  - name: used_bookings
-    description: 'Total number of used bookings generated by a venue'
-  - name: used_individual_bookings
-  - name: used_collective_bookings
-  - name: individual_theoretic_revenue
-  - name: individual_real_revenue
-  - name: collective_theoretic_revenue
-  - name: collective_real_revenue
-  - name: first_individual_offer_creation_date
-    description: First offer creation date on an individual offer
-  - name: last_individual_offer_creation_date
-    description: Last offer creation date on an individual offer
-  - name: first_collective_offer_creation_date
-    description: First offer creation date on a collective offer
-  - name: last_collective_offer_creation_date
-    description: Last offer creation date on a collective offer
-  - name: first_offer_creation_date
-    description: First offer creation date
-  - name: last_offer_creation_date
-    description: Last offer creation date
-  - name: first_booking_date
-    description: First booking date
-  - name: last_booking_date
-    description: Last booking date
-  - name: individual_offers_created
-  - name: collective_offers_created
-  - name: total_offers_created
-    description: Total number of created offers by the venue
-  - name: venue_first_bookable_offer_date
-  - name: venue_last_bookable_offer_date
-  - name: venue_bookable_individual_offer_cnt
-  - name: venue_bookable_collective_offer_cnt
-  - name: venue_bookable_offer_cnt
-  - name: theoretic_revenue
-  - name: real_revenue
-  - name: venue_humanized_id
-  - name: venue_backofficev3_link
-  - name: venue_region_name
-  - name: venue_pc_pro_link
-  - name: venue_targeted_audience
-  - name: banner_url
-  - name: venue_description
-  - name: venue_withdrawal_details
-  - name: venue_contact_phone_number
-  - name: venue_contact_email
-  - name: venue_contact_website
+              error_if: ">1000000"
+    - name: venue_public_name
+    - name: venue_name
+    - name: venue_booking_email
+    - name: venue_address
+    - name: venue_latitude
+    - name: venue_longitude
+    - name: venue_department_code
+    - name: venue_postal_code
+    - name: venue_city
+    - name: venue_siret
+    - name: venue_is_virtual
+    - name: venue_is_permanent
+    - name: venue_managing_offerer_id
+    - name: venue_creation_date
+    - name: offerer_name
+    - name: offerer_validation_status
+      description: "Status of validation of an offerer"      
+    - name : venue_type_label
+      description: Type of venue
+    - name: venue_label
+    - name: total_bookings
+      description: 'Total number of bookings generated by a venue'
+    - name: non_cancelled_individual_bookings
+    - name: non_cancelled_collective_bookings
+    - name: first_individual_booking_date
+      description: First booking date on an individual offer
+    - name: last_individual_booking_date
+      description: Last booking date on an individual offer
+    - name: first_collective_booking_date
+      description: First booking date on a collective offer
+    - name: last_collective_booking_date
+      description: Last booking date on a collective offer
+    - name: non_cancelled_bookings
+      description: Number of non cancelled bookings by the venue.
+    - name: used_bookings
+      description: 'Total number of used bookings generated by a venue'
+    - name: used_individual_bookings
+    - name: used_collective_bookings
+    - name: individual_theoretic_revenue
+    - name: individual_real_revenue
+    - name: collective_theoretic_revenue
+    - name: collective_real_revenue
+    - name: first_individual_offer_creation_date
+      description: First offer creation date on an individual offer
+    - name: last_individual_offer_creation_date
+      description: Last offer creation date on an individual offer
+    - name: first_collective_offer_creation_date
+      description: First offer creation date on a collective offer
+    - name: last_collective_offer_creation_date
+      description: Last offer creation date on a collective offer
+    - name: first_offer_creation_date
+      description: First offer creation date
+    - name: last_offer_creation_date
+      description: Last offer creation date
+    - name: first_booking_date
+      description: First booking date
+    - name: last_booking_date
+      description: Last booking date
+    - name: individual_offers_created
+    - name: collective_offers_created
+    - name: total_offers_created
+      description: Total number of created offers by the venue
+    - name: venue_first_bookable_offer_date
+    - name: venue_last_bookable_offer_date
+    - name: venue_bookable_individual_offer_cnt
+    - name: venue_bookable_collective_offer_cnt
+    - name: venue_bookable_offer_cnt
+    - name: theoretic_revenue
+    - name: real_revenue
+    - name: venue_humanized_id
+    - name: venue_backofficev3_link
+    - name: venue_region_name
+    - name: venue_pc_pro_link
+    - name: venue_targeted_audience
+    - name: banner_url
+    - name: venue_description
+    - name: venue_withdrawal_details
+    - name: venue_contact_phone_number
+    - name: venue_contact_email
+    - name: venue_contact_website
       
 
 exposures:
