+++ orchestration/dags/jobs/dbt/dbt_dag_run.py
@@ -10,7 +10,7 @@ from airflow.models import Param
 from airflow.operators.python import BranchPythonOperator
 from common.alerts import task_fail_slack_alert
 from common.utils import get_airflow_schedule, waiting_operator
-from common.dbt.utils import rebuild_manifest
+from common.dbt.utils import rebuild_manifest, load_manifest
 
 from common import macros
 from common.config import (
@@ -81,26 +81,89 @@ end = DummyOperator(task_id="end", dag=dag, trigger_rule="all_success")
 model_op_dict = {}
 test_op_dict = {}
 
-simplified_manifest = rebuild_manifest(f"/{PATH_TO_DBT_TARGET}")
+# PATH_TO_DBT_TARGET='./target'
+# simplified_manifest = rebuild_manifest(f"/{PATH_TO_DBT_TARGET}")
+manifest = load_manifest(f"{PATH_TO_DBT_TARGET}")
 
+dbt_models = [
+    node
+    for node in manifest["nodes"].keys()
+    if (
+        manifest["nodes"][node]["resource_type"] == "model"
+        and manifest["nodes"][node]["package_name"] == "data_gcp_dbt"
+    )
+]
+dbt_crit_tests = [
+    node
+    for node in manifest["nodes"].keys()
+    if (
+        manifest["nodes"].get(node).get("resource_type") == "test"
+        and manifest["nodes"].get(node).get("package_name") == "data_gcp_dbt"
+    )
+    and manifest["nodes"][node]["config"].get("severity", "warn").lower() == "error"
+]
+models_with_dependencies = [
+    node for node in manifest["child_map"].keys() if node in dbt_models
+]
+models_with_crit_test_dependencies = [
+    manifest["nodes"][node].get("attached_node") for node in dbt_crit_tests
+]
+crit_test_parents = {
+    manifest["nodes"][test].get("attached_node", None): [
+        parent
+        for parent in set(manifest["parent_map"][test]).intersection(set(dbt_models))
+    ]
+    for test in dbt_crit_tests
+}
+models_with_dependencies = [
+    node
+    for node in manifest["child_map"].keys()
+    if (node in dbt_models)
+    and (
+        manifest["nodes"][node]["resource_type"] == "model"
+        and manifest["nodes"][node]["package_name"] == "data_gcp_dbt"
+    )
+]
 
+# first create test operators and hide them in a group
+with TaskGroup(group_id="critical_tests", dag=dag) as crit_test_group:
+    for model_node in dbt_models:
+        full_ref_str = " --full-refresh" if not "{{ params.full_refresh }}" else ""
+        model_data = manifest["nodes"][model_node]
+        if model_node in models_with_crit_test_dependencies:
+            test_op_dict[model_node] = BashOperator(
+                task_id=model_data["alias"] + "_tests",
+                bash_command=f"bash {PATH_TO_DBT_PROJECT}/scripts/dbt_test.sh ",
+                env={
+                    "GLOBAL_CLI_FLAGS": "{{ params.GLOBAL_CLI_FLAGS }}",
+                    "target": "{{ params.target }}",
+                    "model": f"""{model_data['alias']} --exclude "config.severity:warn""",
+                    "full_ref_str": full_ref_str,
+                    "PATH_TO_DBT_TARGET": PATH_TO_DBT_TARGET,
+                },
+                append_env=True,
+                cwd=PATH_TO_DBT_PROJECT,
+                dag=dag,
+            )
+
+# loop over models
 with TaskGroup(group_id="data_transformation", dag=dag) as data_transfo:
-    full_ref_str = " --full-refresh" if not "{{ params.full_refresh }}" else ""
-    # models task group
-    for model_node, model_data in simplified_manifest.items():
-        if model_data["resource_type"] == "model":
-            crit_tests_list = model_data["model_tests"].get("error", [])
-            with TaskGroup(
-                group_id=f'{model_data["model_alias"]}_tasks', dag=dag
-            ) as model_tasks:
-                # models
-                model_op = BashOperator(
-                    task_id=model_data["model_alias"],
+    with TaskGroup(group_id="applicative_tables", dag=dag) as applicative:
+        for model_node in dbt_models:
+            # hide numerous applicative tables
+            if "applicative" in manifest["nodes"][model_node]["alias"]:
+                full_ref_str = (
+                    " --full-refresh" if not "{{ params.full_refresh }}" else ""
+                )
+                model_data = manifest["nodes"][model_node]
+                # with TaskGroup(group_id=f"{model_data['alias']}_tasks", dag=dag):
+                model_op_dict[model_node] = BashOperator(
+                    task_id=model_data["alias"],
                     bash_command=f"bash {PATH_TO_DBT_PROJECT}/scripts/dbt_run.sh ",
                     env={
                         "GLOBAL_CLI_FLAGS": "{{ params.GLOBAL_CLI_FLAGS }}",
                         "target": "{{ params.target }}",
-                        "model": f"{model_data['model_alias']}",
+                        "model": f"{model_data['alias']}",
                         "full_ref_str": full_ref_str,
                         "PATH_TO_DBT_TARGET": PATH_TO_DBT_TARGET,
                     },
@@ -108,71 +171,57 @@ with TaskGroup(group_id="data_transformation", dag=dag) as data_transfo:
                     cwd=PATH_TO_DBT_PROJECT,
                     dag=dag,
                 )
-                model_op_dict[model_data["model_alias"]] = model_op
-                # critical tests task subgroup
-                if len(crit_tests_list) > 0:
-                    with TaskGroup(
-                        group_id=f'{model_data["model_alias"]}_critical_tests', dag=dag
-                    ) as crit_tests_task:
-                        dbt_test_tasks = [
-                            BashOperator(
-                                task_id=test["test_alias"],
-                                bash_command=f"bash {PATH_TO_DBT_PROJECT}/scripts/dbt_run.sh "
-                                if test["test_type"] == "generic"
-                                else f"bash {PATH_TO_DBT_PROJECT}/scripts/dbt_test_model.sh ",
-                                env={
-                                    "GLOBAL_CLI_FLAGS": "{{ params.GLOBAL_CLI_FLAGS }}",
-                                    "target": "{{ params.target }}",
-                                    "model": f"{model_data['model_alias']}",
-                                    "full_ref_str": full_ref_str,
-                                    "PATH_TO_DBT_TARGET": PATH_TO_DBT_TARGET,
-                                },
-                                append_env=True,
-                                cwd=PATH_TO_DBT_PROJECT,
-                                dag=dag,
-                            )
-                            for test in crit_tests_list
-                            if not test["test_alias"].endswith(
-                                f'ref_{model_data["model_alias"]}_'
-                            )
-                        ]
-                        if len(dbt_test_tasks) > 0:
-                            model_op >> crit_tests_task
-                    for i, test in enumerate(crit_tests_list):
+                # create dependencies between tests and their attached model
+                if model_node in models_with_crit_test_dependencies:
+                    model_op_dict[model_node] >> test_op_dict[model_node]
 
-                        if not test["test_alias"].endswith(
-                            f'ref_{model_data["model_alias"]}_'
-                        ):
-                            if test["test_alias"] not in test_op_dict.keys():
-                                test_op_dict[test["test_alias"]] = {
-                                    "parent_model": [model_data["model_alias"]],
-                                    "test_op": dbt_test_tasks[i],
-                                }
-                            else:
-                                test_op_dict[test["test_alias"]]["parent_model"] += [
-                                    model_data["model_alias"]
-                                ]
-                simplified_manifest[model_node]["redirect_dep"] = model_tasks
+    for model_node in dbt_models:
+        if "applicative" not in manifest["nodes"][model_node]["alias"]:
+            full_ref_str = " --full-refresh" if not "{{ params.full_refresh }}" else ""
+            model_data = manifest["nodes"][model_node]
+            # with TaskGroup(group_id=f"{model_data['alias']}_tasks", dag=dag):
+            model_op_dict[model_node] = BashOperator(
+                task_id=model_data["alias"],
+                bash_command=f"bash {PATH_TO_DBT_PROJECT}/scripts/dbt_run.sh ",
+                env={
+                    "GLOBAL_CLI_FLAGS": "{{ params.GLOBAL_CLI_FLAGS }}",
+                    "target": "{{ params.target }}",
+                    "model": f"{model_data['alias']}",
+                    "full_ref_str": full_ref_str,
+                    "PATH_TO_DBT_TARGET": PATH_TO_DBT_TARGET,
+                },
+                append_env=True,
+                cwd=PATH_TO_DBT_PROJECT,
+                dag=dag,
+            )
+            # create dependencies between tests and their attached model
+            if model_node in models_with_crit_test_dependencies:
+                model_op_dict[model_node] >> test_op_dict[model_node]
+
+    # set up models ascendencies
+    for model_node in dbt_models:
+        full_ref_str = " --full-refresh" if not "{{ params.full_refresh }}" else ""
+        model_data = manifest["nodes"][model_node]
+        children = tuple(
+            [
+                model_op_dict[child]
+                for child in manifest["child_map"][model_node]
+                if child in dbt_models
+            ]
+        )
+        # replace model ascendency by test ascendency when needed
+        if model_node in models_with_crit_test_dependencies:
+            test_op_dict[model_node] >> (children if len(children) > 0 else end)
+        else:
+            model_op_dict[model_node] >> (children if len(children) > 0 else end)
 
 
-# models' task groups dependencies
-for node, values in simplified_manifest.items():
-    if values["resource_type"] == "model":
-        for upstream_node in simplified_manifest[node]["depends_on_node"]:
-            if upstream_node is not None:
-                if upstream_node.startswith("model."):
-                    try:
-                        (
-                            simplified_manifest[upstream_node]["redirect_dep"]
-                            >> simplified_manifest[node]["redirect_dep"]
-                        )
-                    except:
-                        pass
+# test's cross dependencies management
+for test, parents in crit_test_parents.items():
+    for p in parents:
+        try:
+            model_op_dict[p] >> test_op_dict[test]
+        except KeyError:
+            pass
 
-# tests' cross dependencies management
-for test_alias, details in test_op_dict.items():
-    for parent in details["parent_model"]:
-        model_op_dict[parent] >> details["test_op"]
-
-
-start >> branching >> [shunt, wait4init] >> join >> data_transfo >> end
+start >> branching >> [shunt, wait4init] >> join >> data_transfo
